move domains list to matcher namespace
